feat(grpc): move proto files to `iota-grpc-types`
