+++ crates/iota-grpc-api/Cargo.toml
@@ -13,9 +13,7 @@ workspace = true
 # external dependencies
 anyhow.workspace = true
 async-stream = "0.3"
-bcs.workspace = true
 futures.workspace = true
-prost.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 tokio = { workspace = true, features = ["rt-multi-thread"] }
@@ -31,9 +29,5 @@ iota-json-rpc-types.workspace = true
 iota-types.workspace = true
 move-core-types.workspace = true
 
-[build-dependencies]
-tonic-build.workspace = true
-
 [dev-dependencies]
-iota-config.workspace = true
 test-cluster.workspace = true
