+++ crates/iota-grpc-types/proto/iota/grpc/v0/common.proto
@@ -3,7 +3,7 @@
 
 syntax = "proto3";
 
-package iota.grpc.common;
+package iota.grpc.v0.common;
 
 // BCS-serialized data container
 message BcsData {
