+++ crates/iota-grpc-types/src/lib.rs
@@ -7,78 +7,26 @@
 //! for future core type evolution. When core types themselves
 //! need versioning, these wrappers will evolve naturally.
 
-use serde::{Deserialize, Serialize};
-
-/// Forward-compatible versioned checkpoint data for gRPC streaming.
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub enum CheckpointData {
-    V1(iota_types::full_checkpoint_content::CheckpointData),
-}
-
-/// Forward-compatible versioned checkpoint summary for gRPC streaming.
-#[derive(Serialize, Deserialize, Clone, Debug)]
-pub enum CertifiedCheckpointSummary {
-    V1(iota_types::messages_checkpoint::CertifiedCheckpointSummary),
-}
-
-impl From<iota_types::full_checkpoint_content::CheckpointData> for CheckpointData {
-    fn from(data: iota_types::full_checkpoint_content::CheckpointData) -> Self {
-        Self::V1(data)
-    }
-}
-
-impl From<iota_types::messages_checkpoint::CertifiedCheckpointSummary>
-    for CertifiedCheckpointSummary
-{
-    fn from(summary: iota_types::messages_checkpoint::CertifiedCheckpointSummary) -> Self {
-        Self::V1(summary)
-    }
-}
-
-impl CheckpointData {
-    /// Extract the V1 checkpoint data, returning None for unknown versions
-    pub fn into_v1(self) -> Option<iota_types::full_checkpoint_content::CheckpointData> {
-        match self {
-            Self::V1(data) => Some(data),
-        }
-    }
-
-    /// Get a reference to the V1 checkpoint data, returning None for unknown
-    /// versions
-    pub fn as_v1(&self) -> Option<&iota_types::full_checkpoint_content::CheckpointData> {
-        match self {
-            Self::V1(data) => Some(data),
-        }
-    }
-
-    /// Get the sequence number regardless of version
-    pub fn sequence_number(&self) -> u64 {
-        match self {
-            Self::V1(data) => data.checkpoint_summary.sequence_number,
+mod proto {
+    pub mod iota {
+        pub mod grpc {
+            pub mod v0 {
+                pub mod common {
+                    tonic::include_proto!("iota.grpc.v0.common");
+                }
+                pub mod checkpoints {
+                    tonic::include_proto!("iota.grpc.v0.checkpoints");
+                }
+                pub mod events {
+                    tonic::include_proto!("iota.grpc.v0.events");
+                }
+            }
         }
     }
 }
 
-impl CertifiedCheckpointSummary {
-    /// Extract the V1 checkpoint summary, returning None for unknown versions
-    pub fn into_v1(self) -> Option<iota_types::messages_checkpoint::CertifiedCheckpointSummary> {
-        match self {
-            Self::V1(summary) => Some(summary),
-        }
-    }
+pub use proto::iota::grpc::v0;
 
-    /// Get a reference to the V1 checkpoint summary, returning None for unknown
-    /// versions
-    pub fn as_v1(&self) -> Option<&iota_types::messages_checkpoint::CertifiedCheckpointSummary> {
-        match self {
-            Self::V1(summary) => Some(summary),
-        }
-    }
-
-    /// Get the sequence number regardless of version
-    pub fn sequence_number(&self) -> u64 {
-        match self {
-            Self::V1(summary) => summary.data().sequence_number,
-        }
-    }
-}
+pub mod bcs;
+pub mod checkpoints;
+pub mod events;
